Properly handle short reads
If a short read from the random device occurred, the index into the data
buffer was not properly updated. While the correct number of bytes would
be read, they would always be written to the start of the data buffer
instead of their proper position. Therefore, instead of proper random
data, the last part of the data buffer would be uninitialized (i.e. not
random) memory.

Fix it by properly updating the index when a short read occurs.